Splits up mods normalizer.
closes #2050
